requirements.txt
aiodns==1.1.1
aiohttp==3.5.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
azure-common==1.1.16
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
cchardet==2.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
idna==2.8
multidict==4.5.2
pycares==2.4.0
pycparser==2.19
python-dateutil==2.7.5
requests==2.21.0
six==1.12.0
urllib3==1.24.1
yarl==1.3.0
